pyproject.toml
[project]
name = "Crawl4AIServer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "crawl4ai==0.3.8",
    "fake-http-header>=0.3.5",
    "fake-useragent>=2.1.0",
    "fastapi>=0.115.12",
    "grpcio>=1.71.0",
    "grpcio-tools>=1.71.0",
    "playwright>=1.47.0",
    "playwright-stealth>=1.0.6",
    "uvicorn>=0.34.0",
]
